March's Updates

+++ Stores/views.py
@@ -13,15 +13,17 @@ from django.http import JsonResponse
 @permission_classes([AllowAny])
 def edit_store(request):
 
-    if not request.data['storeID'] or not request.data['storeIP'] or not request.data['storeNameEN'] or not request.data['storeNumR']:
-      return JsonResponse({"error": "Failed successfully, Please fill all the required fields!"})
-      
+    if not request.data['storeID'] or not request.data['storeNameEN'] or not request.data['storeNumR']:
+      return JsonResponse({"error": "Failed successfully, Please fill all the required fields!"},status=500)
+    print(request.data)
+
     connection, cursor = get_sql_anywhere_connection()
     try:
         store_id = request.data['storeID']
+        org_id = request.data['org_id']
 
-        query_fetch = f"SELECT * FROM \"dba\".\"Stores\" WHERE \"Store\" = ?"
-        cursor.execute(query_fetch, [store_id])
+        query_fetch = f"SELECT * FROM \"dba\".\"Stores\" WHERE \"Store\" = ? AND \"Org_ID\" = ? "
+        cursor.execute(query_fetch, [store_id , org_id])
         existing_data = cursor.fetchone()
         print(request.data)
         if existing_data:
@@ -31,19 +33,21 @@ def edit_store(request):
                 'store_name': request.data['storeNameEN'],
                 'store_name_ar': request.data['storeNameAR'],
                 # 'RegKey': request.data['storeRegKey'],
+                'Lat': request.data['long'],
+                'Lan': request.data['lat'],
                 'number_range': request.data['storeNumR'],
                 'region_id': request.data['storeRegion'],
                 'delivery': request.data['storeDel'],
-                'isTouristic': request.data['storeTouristic'],
+                'isTouristic':request.data['storeTouristic'],
+                'status': request.data['storeStat'],
                 'NetworkType': request.data['storeNetType'],
                 'store_location': request.data['storeLoc'],
-                'status': request.data['storeStat'],
             }
 
             update_columns = ', '.join([f'"{col}" = ?' for col in store_data.keys()])
-            query_update = f"UPDATE \"dba\".\"Stores\" SET {update_columns} WHERE \"Store\" = ?"
+            query_update = f"UPDATE \"dba\".\"Stores\" SET {update_columns} WHERE \"Store\" = ? AND \"Org_ID\" = ? "
 
-            cursor.execute(query_update, list(store_data.values()) + [store_id])
+            cursor.execute(query_update, list(store_data.values()) + [store_id , org_id])
             connection.commit()
 
             return JsonResponse({"message": "Store updated successfully!"}, status=200)
@@ -51,6 +55,7 @@ def edit_store(request):
             return JsonResponse({"error": f"Store with ID {store_id} not found."}, status=404)
 
     except Exception as e:
+        print(str(e))
         return JsonResponse({"error": f"Failed to update store: {str(e)}"}, status=500)
 
     finally:
@@ -64,8 +69,10 @@ def delete_store(request):
     try:
         connection, cursor = get_sql_anywhere_connection()
         store_id = request.data['Store']
+        org_id = request.data['org_id']
+    
         print(store_id)
-        cursor.execute(f"DELETE FROM \"dba\".\"Stores\" WHERE \"Store\" = ?", [store_id])
+        cursor.execute(f"DELETE FROM \"dba\".\"Stores\" WHERE \"Store\" = ? AND \"Org_ID\" = ?", [store_id, org_id])
         
         connection.commit()  
 
@@ -83,11 +90,12 @@ from django.http import JsonResponse
 def add_store(request):
 
     connection, cursor = get_sql_anywhere_connection()
-
+   
     try:
-        if not request.data['storeIP'] or not request.data['storeNameEN'] or not request.data['storeNumR']:
+        if not request.data['storeNameEN'] or not request.data['storeNumR'] or not request.data['org_id']:
             return JsonResponse({"error": "Failed successfully, Please fill all the required fields!"})
-
+        org_id = request.data['org_id']
+        print(request.data)
         store_data = {
             'store': request.data['storeID'],
             'ip': request.data['storeIP'],
@@ -96,10 +104,13 @@ def add_store(request):
             'number_range': request.data['storeNumR'],
             'region_id': request.data['storeRegion'],
             'delivery': request.data['storeDel'],
-            'isTouristic': request.data['storeTouristic'],
+            'isTouristic':1 if request.data['storeTouristic'] == "true" else 0,
+            'status': request.data['storeStat'],
             'NetworkType': request.data['storeNetType'],
             'store_location': request.data['storeLoc'],
-            'status': request.data['storeStat'],
+            'Lat': 0 if request.data['long'] == '' else float(request.data['long']),
+            'Lan': 0 if request.data['lat'] == '' else float(request.data['lat']),
+            'Org_ID' : org_id
         }
 
         query = f"""
@@ -124,14 +135,20 @@ def add_store(request):
 
 from django.http import JsonResponse
 
-@api_view(['GET'])
+@api_view(['GET', 'PUT'])
 @permission_classes([AllowAny])
 def fetch_stores(request):
     connection, cursor = get_sql_anywhere_connection()
 
     try:
-        query = "SELECT * FROM \"dba\".\"Stores\""
-        cursor.execute(query)
+        org_id = request.data['org_id']
+        # query = "SELECT * FROM \"dba\".\"Stores\" "
+        if 'active' in request.data:
+            cursor.execute(f"SELECT * FROM \"dba\".\"Stores\" WHERE \"Org_ID\" = ? AND \"Status\" = 1", [org_id])
+        else:
+            cursor.execute(f"SELECT * FROM \"dba\".\"Stores\" WHERE \"Org_ID\" = ?", [org_id])
+
+        # cursor.execute(query)
 
         rows = cursor.fetchall()
